requirements.txt
asn1crypto==1.3.0
boto3==1.10.50
botocore==1.13.50
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.5
Django==2.2.5
django-environ==0.4.5
django-storages==1.8
django-username-email==2.2.4
docutils==0.15.2
google-api-python-client==1.4.1
gunicorn==19.9.0
httplib2==0.17.0
idna==2.8
jmespath==0.9.4
mock==4.0.1
oauth2client==4.1.3
Pillow==7.0.0
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
responses==0.10.9
rsa==4.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
